+++ src/enc/mod.rs
@@ -66,5 +66,5 @@ pub(crate) fn encode_option_variant<E: Encoder, T>(
 /// Encodes the length of any slice, container, etc into the given encoder
 #[inline]
 pub(crate) fn encode_slice_len<E: Encoder>(encoder: E, len: usize) -> Result<(), EncodeError> {
-    (len as u64).encode(encoder)
+    (len as u32).encode(encoder)
 }
